+++ TestStack.Dossier/TestStack.Dossier.csproj
@@ -52,6 +52,9 @@
   <ItemGroup>
     <Compile Include="AnonymousValueFixture.cs" />
     <Compile Include="Builder.cs" />
+    <Compile Include="DataSources\Pick.cs" />
+    <Compile Include="DataSources\RandomItemSource.cs" />
+    <Compile Include="DataSources\RepeatingSequenceSource.cs" />
     <Compile Include="ITestDataBuilder.cs" />
     <Compile Include="DataSources\Dictionaries\Cache.cs" />
     <Compile Include="DataSources\Dictionaries\IDictionaryRepository.cs" />
@@ -94,10 +97,6 @@
     <Compile Include="NullingExpandoObject.cs" />
     <Compile Include="Factories\IFactory.cs" />
     <Compile Include="PathExpressionVisitor.cs" />
-    <Compile Include="Picking\ItemPicker.cs" />
-    <Compile Include="Picking\Pick.cs" />
-    <Compile Include="Picking\RandomItemPicker.cs" />
-    <Compile Include="Picking\RepeatingSequenceItemPicker.cs" />
     <Compile Include="Reflector.cs" />
     <Compile Include="Suppliers\DefaultEmailValueSupplier.cs" />
     <Compile Include="Suppliers\DefaultLastNameValueSupplier.cs" />
@@ -182,6 +181,9 @@
     <EmbeddedResource Include="DataSources\Dictionaries\Resources\PersonRace.txt" />
     <EmbeddedResource Include="DataSources\Dictionaries\Resources\PersonUsername.txt" />
     <EmbeddedResource Include="DataSources\Dictionaries\Resources\ShirtSize.txt" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Picking\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
